pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "superrepo"
version = "0.1.0"
description = "AI + Automation monorepo"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name="Lord Nibbler"}]

dependencies = [
    "fastapi>=0.112.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "httpx>=0.27.0",
    "loguru>=0.7.2",
    "rich>=13.7.0",
    "typer>=0.12.3",
    "python-multipart>=0.0.9",
    "boto3>=1.34.0",
    "pandas>=2.2.2",
    "numpy>=1.26.4",
    "scikit-learn>=1.5.0",
    "faiss-cpu>=1.8.0",
    "sentence-transformers>=3.0.1",
    "opencv-python>=4.10.0.84",
    "pytesseract>=0.3.13",
    "python-docx>=1.1.2",
    "markdown-it-py>=3.0.0",
    "pdfminer.six>=20231228",
]

[project.optional-dependencies]
all = []
dev = ["pytest>=8.2.0", "black>=24.4.2", "ruff>=0.5.5"]

[tool.ruff]
line-length = 100

[tool.black]
line-length = 100

[tool.setuptools.packages.find]
where = [""]
exclude = ["web*", "apps_script*"]
